wlan: Acquire lock to avoid double free in function freeSmeCmdList.
The function sme_close->freeSmeCmdList gets called in hdd module_exit
context , hdd module_init context ( In case the vos_open fails ) and the
VosWDThread context.While the function is in the process of freeing the
buffer pMac->sme.pSmeCmdBufAddr in either of the context
hdd_module_exit/hdd_module_init gets pre-empted by
VosWDThread ( if SSR kicks in) then it could lead to double free issue.

To address this issue we need to protect it with the spinlocks.

Change-Id: I08ff847f0f3b5c4c8e7187687f68ae9b0ad0f530
CR-Fixed: 426181