wlan: LFR does not bgscan on channel list from cfg.ini
1. In the CSR roaming code, We roll back to cfg.ini channels only when
occupied
channels list is 0. but when we complete the first UI Scan and find a
matching AP,
this occupied channel list is non-zero. hence we won't take the cfg.ini
route.
Fix is to build occupied channels list only if channel list in cfg.ini
(gNeighborScanChannelList) is 0.

2. After the Cl 657615, simple_strtoul kernel API is replaced by
kstrtoul, but the earlier
API used to take care of string buffer with delimiter, but kstrtoul does
not take care
of this and we always see this function returns -ve number.
Fix is to use the other block of code in the same function to use
strpbrk.

Change-Id: I1a1c8ae35a68054948cfbd8ce83f53d439d5c5eb
CR-Fixed: 429923