wlan: Modify the split scan logic to scan more than one channel at a time
Modify the split scan logic to split the full scan into operations which
scan more than one channel at a time. The number of channels that are
scanned in each of the individual scan operations is configurable
through the gNumChanCombinedConc parameter from the cfg.ini file. The
default value of gNumChanCombinedConc is currently set to 1.

Change-Id: Ie5b39dcde3a6cae293cfb9350ae9102a61a84a63
CR-Fixed: 429425