wlan: Fix for STA Scan list goes empty at P2P GO side.
if gScanAgingTime set to 0 and if aging timer is running,
any scan list which is stored for less than 60 seconds(default)
will get cleared. In case of GO mode default value was getting
updated with 0.

Added a fix to set and return gScanAgingTime to initial
configauration in csrGetConfigParam function.

CRs-Fixed: 441439

Change-Id: I5eeb860e5890ae23166cd2676c131815c76f651f