wlan: Clear preauth failed list in CFG_CHAN_LIST_SCAN state
Currently we are only clearing the preauth failed list upon re-entering
INIT state. This means that AP that have failed pre-auth (perhaps for
temporary reasons) are not being reconsidered in a timely manner.
This change removes APs from the failed preauth list when the scan
channel list is recomputed.  For exampe, after a series of scans have
failed to find a candidate.

Change-Id: Ib048fb7566b876c3e85df2259521f170d60f2484
CRs-fixed: 452519