wlan: Avoid sscanf overflow in hdd_parse_send_action_frame_data()
Static source code analysis identified an issue in
hdd_parse_send_action_frame_data() where sscanf() is being called with
an 's' conversion specifier without a width.  This can lead to a
buffer overflow if the parsed string exceeds the size of the buffer.
Add a field width to the 's' conversion specifier to prevent buffer
overflow.

Change-Id: I41d740dec7e0d9e758871d91ca25a8e38611f5a6
CRs-fixed: 470188