wlan: Update csrRoamGetConnectProfile() semantics for IBSS
csrRoamGetConnectProfile() currently returns a profile only if the
session is in a "connected" state.  However Android HDD needs to be
able to retrieve the profile even if the session is in a "disconnected
IBSS" state.  In this state the profile is valid, so update
csrRoamGetConnectProfile() to retrieve the profile when the session is
in any IBSS state.

Change-Id: Ibb90b260035a8f79a72946f14638893e7c6484f4
CRs-fixed: 478476