MAC:No need to set the gLimRspReqd for stats requests.
The function limProcessNormalHddMsg(pMac, limMsg, true) is being
called with fRspReqd as true. Based on this argument while
processing the stats request the PE is resetting the gLimRspReqd
flag. Since these stats requests are not queued to smeCmdActiveList
this might cause issue for the commands which are already in
smeCmdActiveList.

Change-Id: I0e939a4d01cac003f5c45ec4c3170709dc31f885
CRs-fixed: 500914