wlan : rrm : Passive Scan with Beacon Report Request
When AP sends beacon report request for passive scan, the current
implemnetation triggers active scan. Change is made to make STA go
with passive scan in this case. A new field is added in the scan
request command to differentiate if the scan request is triggered
from RRM beacon report request.

Change-Id: Ib1191dec6b0d9ebc8e52f04bc8bbee397bf98dc8
CRs-fixed: 513544