wlan: update sched_scan_results after cfg80211 resumed
wiphy_resume ages the scan results with the time spent
in suspend. If the sched scan results are updated prior
to the resume, the bss ages out further resulting in the
supplicant getting no scan results and hence a failure
in the connection. So update the scan results upon resume.

Change-Id: I571d0d1dbdfdd94ebbc84d2e105f64bb6687c171
CRs-fixed: 526956