scripts: build-all.py: Squashed update from 3.10
* Conditionally revert following commits to remove 64-bit targets:
  "mdm9630: rename msmkrypton to mdm9630" (I0fd861f30bb461ea575084bfe241d8ded1d0c484)
  "scripts: Reorganize build-all for multi arch" (I0bd9a7bff22f9c21371c663aff9e8be5116fcdae)
  "scripts: Use custom defconfig for 64-bit targets" (I2c3840fd668aa9a1dbd304c64484f6735f1aeea1)

- - -

build-all.py: Add support for mpq*_defconfig

mpq8092_defconfig does not match the existing msm pattern
because of not having its name started with msm prefix.
Compile it too.

Change-Id: I1b758e33fa550fdac539a18cc61a6a8348f61f71
Signed-off-by: Srinivas Ramana <sramana@codeaurora.org>

mdm9630: rename msmkrypton to mdm9630

Introduce mdm9630 as the official name for msmkrypton.

Change-Id: I0fd861f30bb461ea575084bfe241d8ded1d0c484
Signed-off-by: Joonwoo Park <joonwoop@codeaurora.org>

scripts: Reorganize build-all for multi arch

Move most of the build work into a Builder class, which is given a
name and a specific defconfig.  If the arch contains arm64, build
using the CROSS_COMPILE64 compiler rather than the CROSS_COMPILE
compiler.

Defconfigs that match the globs in arm64 will have a "-64" appended to
the name, to avoid conflicts when the same target has both a 32-bit
and a 64-bit build.

The main visible change to this is that CROSS_COMPILE now must be set
before invoking this script.  If CROSS_COMPILE64 isn't set, it won't
be willing to build any 64-bit targets.

Change-Id: I0bd9a7bff22f9c21371c663aff9e8be5116fcdae
Signed-off-by: David Brown <davidb@codeaurora.org>

scripts: Keep appending log between build parts

Instead of wiping the log file for each make target, only wipe it
initially, and then continue to append to it.  Instead of closing
between files, just flush.  Although the last use doesn't ever close
the file, it is flushed, and will be closed on exit.

Change-Id: Ic4b38cdad993bc32711f832ce30ca8d8dead0db6
Signed-off-by: David Brown <davidb@codeaurora.org>

scripts: Use custom defconfig for 64-bit targets

Use a different set of defconfig patterns for 64-bit targets.
Currently, msm_defconfig needs to not be built for 32-bit, and is the
only defconfig for the 64-bit target.

Change-Id: I2c3840fd668aa9a1dbd304c64484f6735f1aeea1
Signed-off-by: David Brown <davidb@codeaurora.org>

scripts: Refactor build-all.py's build

Build all consists of a mismash of figuring out variables, running
shutils, and actually executing commands.  In preparation for parallel
builds, change this to build up the commands in advance, and then have
a tracker go through and actually run the commands.

Change-Id: I24215b14e5d045f8269bbdc2af9989ab21b2310b
Signed-off-by: David Brown <davidb@codeaurora.org>

scripts: Build in parallel if requested

Instead of passing -jn directly through to make, and then building
each target in sequence, use the -j and -l options to determine a
number of threads to perform parallel invocations of make in.

The build results are collected, and printed when each build finishes,
which eliminates the interleaving that would otherwise occur between
builds.

Change-Id: Ieef0dcc8adfc72cbe043a3491d4d6f6c23fede31
Signed-off-by: David Brown <davidb@codeaurora.org>

- - -

Change-Id: Ie8247bb30e860e6ba5f5f81878b31c45f9e74d91
Signed-off-by: Caio Oliveira <caiooliveirafarias0@gmail.com>